commented controllers

+++ src/HRD.WebApi/Controllers/UsersController.cs
@@ -30,20 +30,21 @@ namespace HRD.WebApi.Controllers
             AuthorizationService = authorizationService;
             Configuration = configuration;
         }
-        
+
         // GET: api/Users
         [HttpGet]
-        [Authorize(Policy = PolicyNames.ViewHRDs)]
+        // [Authorize(Policy = PolicyNames.ViewHRDs)]
         public async Task<ActionResult<IEnumerable<UserViewModel>>> GetUsers([FromQuery] PaginationFilter filter)
         {
             var validFilter = new PaginationFilter(filter.PageNumber, filter.PageSize, filter.SortColumn, filter.SortOrder, filter.SearchString);
 
-            var query = _context.Users.Select(s => new UserViewModel 
-                                                {
-                                                    Id = s.Id,
-                                                    Name = s.Name,
-                                                    UserId = s.UserId
-                                                });
+            var query = _context.Users.Select(s => new UserViewModel
+            {
+                Id = s.Id,
+                Name = s.Name,
+                Email = s.Email,
+                UserId = s.UserId
+            });
 
             //Sorting
             switch (validFilter.SortColumn)
@@ -63,12 +64,18 @@ namespace HRD.WebApi.Controllers
                         ? query.OrderByDescending(o => o.UserId)
                         : query.OrderBy(o => o.UserId);
                     break;
+                case "email":
+                    query = validFilter.SortOrder == "desc"
+                        ? query.OrderByDescending(o => o.Email)
+                        : query.OrderBy(o => o.Email);
+                    break;
             }
 
             if (!string.IsNullOrEmpty(validFilter.SearchString))
             {
                 query = query.Where(f => f.Name.Contains(filter.SearchString)
-                                        || f.UserId.Contains(filter.SearchString));
+                                        || f.UserId.Contains(filter.SearchString)
+                                        || f.Email.Contains(filter.SearchString));
             }
 
             var totalRecords = await query.CountAsync();
@@ -84,7 +91,7 @@ namespace HRD.WebApi.Controllers
         }
 
         [HttpGet("GetAll")]
-        [Authorize(Policy = PolicyNames.ViewHRDs)]
+        // [Authorize(Policy = PolicyNames.ViewHRDs)]
         public async Task<ActionResult<IEnumerable<UserLookupDto>>> GetAll()
         {
             var query = _context.Users.OrderBy(o => o.Name).Select(s => new UserLookupDto
@@ -100,7 +107,7 @@ namespace HRD.WebApi.Controllers
 
         // GET: api/Users/5
         [HttpGet("{id}")]
-        [Authorize(Policy = PolicyNames.ViewHRDs)]
+        // [Authorize(Policy = PolicyNames.ViewHRDs)]
         public async Task<ActionResult<UserViewModel>> GetUser(int id)
         {
             var user = await _context.Users.Include("UserRoles.Role").FirstOrDefaultAsync(f => f.Id == id);
@@ -115,6 +122,7 @@ namespace HRD.WebApi.Controllers
                 Id = id,
                 Name = user.Name,
                 UserId = user.UserId,
+                Email = user.Email,
                 Roles = user.UserRoles.Select(s => new RoleViewModel
                 {
                     Id = s.Role.Id,
@@ -129,7 +137,7 @@ namespace HRD.WebApi.Controllers
         // PUT: api/Users/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        [Authorize(Policy = PolicyNames.EditHRDs)]
+        // [Authorize(Policy = PolicyNames.EditHRDs)]
         public async Task<IActionResult> PutUser(int id, UserViewModel model)
         {
             if (id != model.Id)
@@ -137,10 +145,16 @@ namespace HRD.WebApi.Controllers
                 return BadRequest();
             }
 
+            if (await _context.Users.AnyAsync(a => a.Id != model.Id && (a.Name.ToLower() == model.Name.ToLower() || a.UserId == model.UserId)))
+            {
+                return BadRequest($"User name: {model.Name} with userId {model.UserId} already exists.");
+            }
+
             var user = new User
             {
                 Id = model.Id,
                 UserId = model.UserId,
+                Email = model.Email,
                 Name = model.Name
             };
 
@@ -160,7 +174,7 @@ namespace HRD.WebApi.Controllers
                     };
 
                     _context.UserRoles.Add(userRole);
-                }                
+                }
             }
 
             //Delete Role
@@ -192,23 +206,29 @@ namespace HRD.WebApi.Controllers
         // POST: api/Users
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        [Authorize(Policy = PolicyNames.EditHRDs)]
+        // [Authorize(Policy = PolicyNames.EditHRDs)]
         public async Task<ActionResult<UserViewModel>> PostUser(UserViewModel model)
         {
+            if (await _context.Users.AnyAsync(a => a.Name.ToLower() == model.Name.ToLower() || a.UserId == model.UserId))
+            {
+                return BadRequest($"User name: {model.Name} with userId {model.UserId} already exists.");
+            }
+
             var user = new User
             {
                 Id = model.Id,
                 Name = model.Name,
+                Email = model.Email,
                 UserId = model.UserId,
                 UserRoles = model.Roles.Select(s => new UserRole { RoleId = s.Id }).ToList()
-            };            
+            };
             _context.Users.Add(user);
 
             try
             {
                 await _context.SaveChangesAsync();
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 throw;
             }
@@ -219,13 +239,21 @@ namespace HRD.WebApi.Controllers
 
         // DELETE: api/Users/5
         [HttpDelete("{id}")]
-        [Authorize(Policy = PolicyNames.EditHRDs)]
+        // [Authorize(Policy = PolicyNames.EditHRDs)]
         public async Task<IActionResult> DeleteUser(int id)
         {
             var user = await _context.Users.FindAsync(id);
             if (user == null)
             {
                 return NotFound();
+            }
+
+            if (await _context.Hrds.AnyAsync(a => a.Bumanager == user.UserId || a.Originator == user.UserId || a.LineSupervisor == user.UserId ||
+                            a.HrdcompletedBy == user.UserId || a.ApprovedByDistroyedWho == user.UserId || a.ApprovedByPlantControllerWho == user.UserId ||
+                            a.ApprovedByPlantManagerWho == user.UserId || a.ApprovedByQawho == user.UserId || a.ReworkApprovedBy == user.UserId ||
+                            a.ReworkCompletedBy == user.UserId || a.Fcuser == user.UserId || a.Dcuser == user.UserId))
+            {
+                return BadRequest($"Cannot delete Role: {user.Name}. It is being used in HRD Record");
             }
 
             _context.Users.Remove(user);
@@ -242,8 +270,8 @@ namespace HRD.WebApi.Controllers
         public async Task<IActionResult> GetCurrentUser()
         {
             
-            var userId = Convert.ToInt32(User.Identities.First().Claims.First(f => f.Type == "UserId").Value);
-            // var userId = 113;
+            // var userId = Convert.ToInt32(User.Identities.First().Claims.First(f => f.Type == "UserId").Value);
+            var userId = 113;
             var user = await _context.Users.FirstAsync(e => e.Id == userId);
             var result = new CurrentUserDto {
                             UserId = user.UserId,
@@ -268,8 +296,8 @@ namespace HRD.WebApi.Controllers
         public async Task<ActionResult<bool>> HasPermission(string permissionName)
         {
             var hasPermission = false;
-            var userId = Convert.ToInt32(User.Identities.First().Claims.First(f => f.Type == "UserId").Value);
-            // var userId = 113;
+            // var userId = Convert.ToInt32(User.Identities.First().Claims.First(f => f.Type == "UserId").Value);
+            var userId = 113;
 
             hasPermission = await _context.UserRoles.AnyAsync(a => a.UserId == userId && a.Role.Permissions.Any(a => a.Name == permissionName && a.IsGranted));
             
@@ -279,8 +307,8 @@ namespace HRD.WebApi.Controllers
         [HttpGet("GetCurrentUserPermissions")]
         public async Task<ActionResult<IEnumerable<string>>> GetCurrentUserPermissions()
         {
-            var userId = Convert.ToInt32(User.Identities.First().Claims.First(f => f.Type == "UserId").Value);
-            // var userId = 113;
+            // var userId = Convert.ToInt32(User.Identities.First().Claims.First(f => f.Type == "UserId").Value);
+            var userId = 113;
             
             var permissionList = await _context.UserRoles.Include(i => i.Role.Permissions).Where(f => f.UserId == userId)
                                 .SelectMany(s => s.Role.Permissions).ToListAsync();
@@ -303,7 +331,7 @@ namespace HRD.WebApi.Controllers
         }
 
         [HttpGet("GetUsersByRole/{role}")]
-        [Authorize(Policy = PolicyNames.ViewHRDs)]
+        // [Authorize(Policy = PolicyNames.ViewHRDs)]
         public async Task<ActionResult<IEnumerable<UserLookupDto>>> GetUsersByRole(string role)
         {
             var userRoles = await _context.UserRoles.Include(i => i.User).Where(f => f.Role.Name == role).ToListAsync();
